third_party/blink/web_tests/virtual/xslt-enabled/http/tests/xsl/xslt/permissions-policy-frame-policy-in-xsl-quirk-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: async_test is not defined
